refactor: 애니메이션 컴포넌트를 LottieAnimation에서 AnimationLottie로 변경 및 이미지 파일로 업데이트
