+++ spark-web/src/pages/Result.tsx
@@ -1,13 +1,13 @@
 import analysis from '@/assets/result/analaysis.json';
 import { ResultInteraction } from '@/domains/Results/components';
 import { NextPageFooter } from '@/shared/components';
-import LottieAnimation from '@/shared/ui/components/AnimationLottie';
+import { AnimationLottie } from '@/shared/ui';
 
 export default function Result() {
   return (
     <div>
       <div className={'relative z-10'}>
-        <LottieAnimation animationData={analysis} height="100%" />
+        <AnimationLottie animationData={analysis} height="100%" />
       </div>
       <div className={'flex justify-center'}>
         <ResultInteraction channelName={'Spark'} />
